% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idle.R
\name{idle_update}
\alias{idle_update}
\title{Watch R Session to Detect If Idle Sessions}
\usage{
idle_update(delay = 15 * 60)
}
\arguments{
\item{delay}{Delay between session polling}
}
\description{
Watch R Session to Detect If Idle Sessions
}
